+++ s/got.ts
@@ -1,8 +1,0 @@
-
-import {is} from "./is.js"
-
-export function got<X>(thing: X | undefined | null): X {
-	if (is.sad(thing)) throw new Error(`got failed`)
-	return thing
-}
-